rework commutable exports and imports
