--- packages/commutable/index.ts
@@ -1,5 +1,0 @@
-/**
- * @module commutable
- */
-
-export * from "./src";
