+++ packages/commutable/src/index.ts
@@ -7,7 +7,12 @@
 //
 
 export * from "./primitives";
+export * from "./structures";
+export * from "./outputs";
+export * from "./cells";
+export * from "./notebook";
 
+/*
 // from structures
 export {
   emptyCodeCell,
@@ -53,3 +58,5 @@ export {
   Notebook,
   ImmutableNotebook
 } from "./notebook";
+
+*/
